refactor: 별점 타입 변경

+++ backend/user/src/main/java/com/a703/user/entity/HistoryEntity.java
@@ -23,7 +23,7 @@ public class HistoryEntity extends BaseEntity{
     private UserEntity reviewee;
 
     private Boolean owner;
-    private Double star;
+    private Integer star;
     private String review;
     private String recordIdx;
 }
